+++ tests/Config.php
@@ -7,8 +7,6 @@
  * Smarty PHPUnit Config
  */
 define('individualFolders', true);
-define('MemCacheEnable', false);
-define('ApcCacheEnable', false);
 define('MysqlCacheEnable', false);
 define('PdoCacheEnable', false);
 define('PdoGzipCacheEnable', false);