+++ tests/UnitTests/TemplateSource/X_Scopes/ScopeTest.php
@@ -291,6 +291,9 @@ class ScopeTest extends PHPUnit_Smarty
                            'no  smarty', $i ++,),);
     }
 
+	/**
+	 * @doesNotPerformAssertions
+	 */
     public function testFunctionScope()
     {
         $this->smarty->assign('scope', 'none');